requirements.txt
coolsms-python-sdk==2.0.3
Django==1.11.2
pytz==2017.2
